Implement backup of board
This patch introduces a basic backup function to Trollolo. It adds
a command `backup` to create a backup of a board identified by its id
and a command `show_backup` to show the content of the backup.

It also adds a command `list_backups` which shows the list of backups
which have been made.

The backups are stored to the directory `~/.trollolo/backups` in JSON
format.