Don't hide exceptions
When the parsing fails for whatever reason, the exception should
be reported to the user. If there are cases where we actually want
to only print an error message and not exit, we should explicitly
catch the exception, which we want to treat specially, e.g. not
`Exception` but the parse error the YAML parser throws when the
syntax of the YAML is incorrect.