Add test for helper script generating chart
This tests the Python helper, which does the actual image generation.
It compares the actual result with a reference file.

The test depends on matplotlib to be installed and configured
correctly, so their results might vary depending on which system
they are run.

Exlude the integration tests from the Travis run to get more stable
and fast results.