Don't overwrite data on first day of sprint
The data of the first day of the sprint is used as starting point,
so it should not be updated on subsequent runs of Trollolo. The
progress of day one is reflected in the data of day two.

This gives a slight assymetry in data, because there is no progress
during the first day, but otherwise the starting point would be lost.
The data is still exact enough, as its granularity and accuracy is
on the order of what happens on a day not less.