Run image comparison tests in defined environment
The integration tests comparing the images generated by the python
script to reference images are run in a Docker environmant now, so
that the results are better reproducible.

The test were fragile because of subtle differences e.g. in how the
anti-aliasing is done which in fact didn't change the results.