Fix calculation of unplanned tasks on day one
The calculation of the unplanned tasks was done in a wrong way by referring to
the number of 'done' in the 'unplanned_tasks' section. However, there's no
'done', but 'total' and 'open' values for unplanned tasks. Therefore the
calculation is now done by subtracting 'open' from 'total'.

An additional yaml file and the resulting plot (png) have been added to test
this case.