Fix issue #44 - set-priorities command fails with cards having '#' in the title
Trollolo has a subset of functions that change data on a trello board. To
do so trollolo generates URLs which can contain data read from the trello
board.
Since those data might contain characters which are invalid in an URI or have
a certain meaning, we need to escape them.