Make `move-backlog` to work without waterline
Having a waterline and a seabed card in the `move-backlog` command was
compulsory. Not every team have this two cards, so this should work
without them too.

Fixes https://github.com/openSUSE/trollolo/issues/107.

The error message when one of this cards was not there was not clear,
but as this error is not error any more this also closes
https://github.com/openSUSE/trollolo/issues/106.