Allow to create an Sprint with a custom number
It should be possible to provide a custom number number for creating a
sprint. As we are currently always updating the sprint with higher
number that also means that we need to be able to modify the sprint for
a concrete Sprint. So, I had implemented a new option `sprint-number`
which can be used for both cases.

Closes https://github.com/openSUSE/trollolo/issues/78