Add support for swimlanes
Swimlanes are used for cards which are not part of the normal sprint
data. They are ignored for the burndown chart and their data is
collected separately.

Swimlane cards are indicated by a user-defined label. The label has
to be on the cards which are in the swimlane and the name of the
label has to be specified in the configuration or the burndown
data YAML file.